feat(token): add token package
