+++ pkg/token/quote_test.go
@@ -1,4 +1,4 @@
-package parser
+package token
 
 import (
 	"testing"
